requirements.txt
requests
pyyaml
pathlib
tqdm
pycrypto